+++ src/main/telemetry/crsf.c
@@ -614,7 +614,7 @@ void crsfSensorEncodeControl(telemetrySensor_t *sensor, sbuf_t *buf)
     const int p = lrintf(mixerGetInput(MIXER_IN_STABILIZED_PITCH) * 1200);
     const int r = lrintf(mixerGetInput(MIXER_IN_STABILIZED_ROLL) * 1200);
     const int y = lrintf(mixerGetInput(MIXER_IN_STABILIZED_YAW) * 800);
-    const int c = lrintf(mixerGetInput(MIXER_IN_STABILIZED_COLLECTIVE) * 1200);
+    const int c = 0;
     sbufWriteU8(buf, ((p >> 8) & 0x0F) | ((r >> 4) & 0xF0));
     sbufWriteU8(buf, (p & 0xFF));
     sbufWriteU8(buf, (r & 0xFF));
@@ -720,7 +720,6 @@ static telemetrySensor_t crsfCustomTelemetrySensors[] =
     TLM_SENSOR(PITCH_CONTROL,           0x1031,   200,  3000,    10,    S16),
     TLM_SENSOR(ROLL_CONTROL,            0x1032,   200,  3000,    10,    S16),
     TLM_SENSOR(YAW_CONTROL,             0x1033,   200,  3000,    10,    S16),
-    TLM_SENSOR(COLLECTIVE_CONTROL,      0x1034,   200,  3000,    10,    S16),
     TLM_SENSOR(THROTTLE_CONTROL,        0x1035,   200,  3000,    10,    S8),
 
     TLM_SENSOR(ESC1_VOLTAGE,            0x1041,   200,  3000,    10,    U16),